Move exposeContextBeans to UrlBasedViewResolver
Prior to this commit, the ability to expose all (or a named subset)
of context beans lived on InternalResourceView and
InternalResourceViewResolver.  This changes moves this functionality
up the hierarchy to AbstractView/UrlBasedViewResolver.

Issue: SPR-8064
