+++ spring-webmvc/src/main/java/org/springframework/web/servlet/view/AbstractView.java
@@ -18,11 +18,15 @@ package org.springframework.web.servlet.view;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
+import java.util.Arrays;
 import java.util.Collections;
+import java.util.HashSet;
 import java.util.LinkedHashMap;
 import java.util.Map;
 import java.util.Properties;
+import java.util.Set;
 import java.util.StringTokenizer;
+
 import javax.servlet.ServletOutputStream;
 import javax.servlet.http.HttpServletRequest;
 import javax.servlet.http.HttpServletResponse;
@@ -30,6 +34,7 @@ import javax.servlet.http.HttpServletResponse;
 import org.springframework.beans.factory.BeanNameAware;
 import org.springframework.http.MediaType;
 import org.springframework.util.CollectionUtils;
+import org.springframework.web.context.support.ContextExposingHttpServletRequest;
 import org.springframework.web.context.support.WebApplicationObjectSupport;
 import org.springframework.web.servlet.View;
 import org.springframework.web.servlet.support.RequestContext;
@@ -73,6 +78,11 @@ public abstract class AbstractView extends WebApplicationObjectSupport implement
 
 	/** Whether or not the view should add path variables in the model */
 	private boolean exposePathVariables = true;
+	
+	private boolean exposeContextBeansAsAttributes = false;
+
+	private Set<String> exposedContextBeanNames;
+	
 
 
 	/**
@@ -248,7 +258,37 @@ public abstract class AbstractView extends WebApplicationObjectSupport implement
 		return this.exposePathVariables;
 	}
 
+	/**
+	 * Set whether to make all Spring beans in the application context accessible
+	 * as request attributes, through lazy checking once an attribute gets accessed.
+	 * <p>This will make all such beans accessible in plain {@code ${...}}
+	 * expressions in a JSP 2.0 page, as well as in JSTL's {@code c:out}
+	 * value expressions.
+	 * <p>Default is "false". Switch this flag on to transparently expose all
+	 * Spring beans in the request attribute namespace.
+	 * <p><b>NOTE:</b> Context beans will override any custom request or session
+	 * attributes of the same name that have been manually added. However, model
+	 * attributes (as explicitly exposed to this view) of the same name will
+	 * always override context beans.
+	 * @see #getRequestToExpose
+	 */
+	public void setExposeContextBeansAsAttributes(boolean exposeContextBeansAsAttributes) {
+		this.exposeContextBeansAsAttributes = exposeContextBeansAsAttributes;
+	}
 
+	/**
+	 * Specify the names of beans in the context which are supposed to be exposed.
+	 * If this is non-null, only the specified beans are eligible for exposure as
+	 * attributes.
+	 * <p>If you'd like to expose all Spring beans in the application context, switch
+	 * the {@link #setExposeContextBeansAsAttributes "exposeContextBeansAsAttributes"}
+	 * flag on but do not list specific bean names for this property.
+	 */
+	public void setExposedContextBeanNames(String[] exposedContextBeanNames) {
+		this.exposedContextBeanNames = new HashSet<String>(Arrays.asList(exposedContextBeanNames));
+	}
+
+	
 	/**
 	 * Prepares the view given the specified model, merging it with static
 	 * attributes and a RequestContext attribute, if necessary.
@@ -264,9 +304,31 @@ public abstract class AbstractView extends WebApplicationObjectSupport implement
 
 		Map<String, Object> mergedModel = createMergedOutputModel(model, request, response);
 		prepareResponse(request, response);
-		renderMergedOutputModel(mergedModel, request, response);
+
+		// Determine which request handle to expose to the RequestDispatcher.
+		HttpServletRequest requestToExpose = getRequestToExpose(request);
+
+		renderMergedOutputModel(mergedModel, requestToExpose, response);
 	}
 
+	
+	/**
+	 * Get the request handle to expose to the RequestDispatcher, i.e. to the view.
+	 * <p>The default implementation wraps the original request for exposure of
+	 * Spring beans as request attributes (if demanded).
+	 * @param originalRequest the original servlet request as provided by the engine
+	 * @return the wrapped request, or the original request if no wrapping is necessary
+	 * @see #setExposeContextBeansAsAttributes
+	 * @see org.springframework.web.context.support.ContextExposingHttpServletRequest
+	 */
+	protected HttpServletRequest getRequestToExpose(HttpServletRequest originalRequest) {
+		if (this.exposeContextBeansAsAttributes || this.exposedContextBeanNames != null) {
+			return new ContextExposingHttpServletRequest(
+					originalRequest, getWebApplicationContext(), this.exposedContextBeanNames);
+		}
+		return originalRequest;
+	}
+	
 	/**
 	 * Creates a combined output Map (never {@code null}) that includes dynamic values and static attributes.
 	 * Dynamic values take precedence over static attributes.
