+++ spring-webmvc/src/main/java/org/springframework/web/servlet/view/InternalResourceViewResolver.java
@@ -52,11 +52,6 @@ public class InternalResourceViewResolver extends UrlBasedViewResolver {
 
 	private Boolean alwaysInclude;
 
-	private Boolean exposeContextBeansAsAttributes;
-
-	private String[] exposedContextBeanNames;
-
-
 	/**
 	 * Sets the default {@link #setViewClass view class} to {@link #requiredViewClass}:
 	 * by default {@link InternalResourceView}, or {@link JstlView} if the JSTL API
@@ -88,42 +83,13 @@ public class InternalResourceViewResolver extends UrlBasedViewResolver {
 	public void setAlwaysInclude(boolean alwaysInclude) {
 		this.alwaysInclude = Boolean.valueOf(alwaysInclude);
 	}
-
-	/**
-	 * Set whether to make all Spring beans in the application context accessible
-	 * as request attributes, through lazy checking once an attribute gets accessed.
-	 * <p>This will make all such beans accessible in plain {@code ${...}}
-	 * expressions in a JSP 2.0 page, as well as in JSTL's {@code c:out}
-	 * value expressions.
-	 * <p>Default is "false".
-	 * @see InternalResourceView#setExposeContextBeansAsAttributes
-	 */
-	public void setExposeContextBeansAsAttributes(boolean exposeContextBeansAsAttributes) {
-		this.exposeContextBeansAsAttributes = exposeContextBeansAsAttributes;
-	}
-
-	/**
-	 * Specify the names of beans in the context which are supposed to be exposed.
-	 * If this is non-null, only the specified beans are eligible for exposure as
-	 * attributes.
-	 * @see InternalResourceView#setExposedContextBeanNames
-	 */
-	public void setExposedContextBeanNames(String[] exposedContextBeanNames) {
-		this.exposedContextBeanNames = exposedContextBeanNames;
-	}
-
+	
 
 	@Override
 	protected AbstractUrlBasedView buildView(String viewName) throws Exception {
 		InternalResourceView view = (InternalResourceView) super.buildView(viewName);
 		if (this.alwaysInclude != null) {
 			view.setAlwaysInclude(this.alwaysInclude);
-		}
-		if (this.exposeContextBeansAsAttributes != null) {
-			view.setExposeContextBeansAsAttributes(this.exposeContextBeansAsAttributes);
-		}
-		if (this.exposedContextBeanNames != null) {
-			view.setExposedContextBeanNames(this.exposedContextBeanNames);
 		}
 		view.setPreventDispatchLoop(true);
 		return view;
